ai-voice-agent-backend/pyproject.toml
[project]
name = "ai-voice-agent-backend"
version = "0.1.0"
description = "AI Voice Agent for medical appointment scheduling using LiveKit, Claude, Deepgram, Cartesia, and Tavus"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[deepgram,cartesia,anthropic,silero,tavus,turn-detector]~=1.3",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
